# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gerenciamento de Consultas",
        "description": "API para gerenciamento de consultas médicas",
        "contact": {
            "email": "seu-email@exemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/medicos": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Lista todos os médicos",
                "description": "Retorna a lista de médicos cadastrados no sistema.",
                "operationId": "7932891d029d6c7dbedd3a9a4d54008f",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Filtrar médicos pelo nome",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de médicos recuperada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de médicos recuperada com sucesso"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Cadastra um novo médico",
                "description": "Cria um novo cadastro de médico no sistema.",
                "operationId": "c4e164db5f38a2ca833b5a56526d2952",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "especialidade"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Dr. João Silva"
                                    },
                                    "especialidade": {
                                        "type": "string",
                                        "example": "Cardiologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Médico cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Médico cadastrado com sucesso"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cidades/{cidade_id}/medicos": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Lista médicos por cidade",
                "description": "Retorna a lista de médicos em uma cidade específica.",
                "operationId": "0fdfe665465924ce4fbb09cce06b4fc0",
                "parameters": [
                    {
                        "name": "cidade_id",
                        "in": "path",
                        "description": "ID da cidade",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de médicos recuperada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de médicos por cidade recuperada com sucesso"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/v1/medicos/{medicoId}/pacientes": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Lista os pacientes de um médico",
                "description": "Retorna a lista de pacientes de um médico específico.",
                "operationId": "6b1073a5732d7971d0f011f51c5ce620",
                "parameters": [
                    {
                        "name": "medicoId",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pacientes recuperada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de pacientes recuperada com sucesso"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}